+++ src/strftime.c
@@ -283,7 +283,7 @@ format_time(char *buf, int max_len, const char *fmt, struct tm *tmval)
 					tmp[2] = '\0';
 					break;
 				}
-				if (pos + strlen(tmp) > max_len)
+				if (pos + strlen(tmp) > (size_t)max_len)
 					return (0);
 				(void) strcpyn(buf + pos, max_len - pos, tmp);
 				pos += strlen(tmp);
